feat(org-details): Linking to org & redirects;
This patch adds links from the `<Directory/>` component to the Dynamic Route at
`org/[id]/page.tsx`. The Organization page is using the Mock Directory data and
if the organization cannot be found in that the data, the page component
triggers a `<NotFound/>` component which renders that page. More info can be
found in the links below.

Dynamic Routes: https://nextjs.org/docs/app/building-your-application/routing/dynamic-routes
NotFound function: https://nextjs.org/docs/app/api-reference/functions/not-found
NotFound file: https://nextjs.org/docs/app/api-reference/file-conventions/not-found
